Moved the http lib into a transport folder 	modified:   Cargo.toml 	modified:   examples/http.rs 	modified:   src/lib.rs 	renamed:    src/http/mod.rs -> src/transport/http/mod.rs 	renamed:    src/http/rest/mock.rs -> src/transport/http/rest/mock.rs 	renamed:    src/http/rest/mod.rs -> src/transport/http/rest/mod.rs
