+++ src/lib.rs
@@ -1,8 +1,8 @@
-pub mod http;
+pub mod transport;
 
 use std::{fmt::Display, hash::Hash};
 
 pub trait Identifiable {
-    type Id: Eq + Hash + Default + Display;
+    type Id: Default + Display + Eq + Hash + Copy;
     fn id(&self) -> Self::Id;
 }
